Reorganized source code
