+++ dbc/dbcmaineditor.cpp
