+++ dbc/dbcmaineditor.h
