+++ dbc/dbcsignaleditor.cpp
