+++ dbc/dbcsignaleditor.h
