+++ mainwindow.h
@@ -9,24 +9,24 @@
 #include "can_structs.h"
 #include "framefileio.h"
 #include "graphingwindow.h"
-#include "frameinfowindow.h"
+#include "re/frameinfowindow.h"
 #include "frameplaybackwindow.h"
-#include "flowviewwindow.h"
+#include "re/flowviewwindow.h"
 #include "framesenderwindow.h"
-#include "filecomparatorwindow.h"
-#include "dbchandler.h"
-#include "dbcmaineditor.h"
+#include "re/filecomparatorwindow.h"
+#include "dbc/dbchandler.h"
+#include "dbc/dbcmaineditor.h"
 #include "mainsettingsdialog.h"
 #include "firmwareuploaderwindow.h"
-#include "discretestatewindow.h"
+#include "re/discretestatewindow.h"
 #include "connectionwindow.h"
 #include "scriptingwindow.h"
-#include "rangestatewindow.h"
-#include "dbcloadsavewindow.h"
-#include "fuzzingwindow.h"
-#include "udsscanwindow.h"
+#include "re/rangestatewindow.h"
+#include "dbc/dbcloadsavewindow.h"
+#include "re/fuzzingwindow.h"
+#include "re/udsscanwindow.h"
 #include "re/sniffer/snifferwindow.h"
-#include "isotp_interpreterwindow.h"
+#include "re/isotp_interpreterwindow.h"
 #include "motorcontrollerconfigwindow.h"
 
 class ConnectionWindow;
