+++ re/discretestatewindow.cpp
