+++ re/filecomparatorwindow.cpp
