+++ re/filecomparatorwindow.h
