+++ re/flowviewwindow.cpp
