+++ re/frameinfowindow.cpp
