+++ re/frameinfowindow.h
