+++ re/rangestatewindow.cpp
