+++ re/rangestatewindow.h
