+++ ui/connectionwindow.ui
