+++ ui/dbcsignaleditor.ui
