+++ ui/filecomparatorwindow.ui
