+++ ui/firmwareuploaderwindow.ui
