+++ ui/frameinfowindow.ui
