+++ ui/graphingwindow.ui
