+++ ui/newgraphdialog.ui
