+++ ui/scriptingwindow.ui
